Create 227. Basic Calculator II.py
https://www.loom.com/share/92945e1654f84f579a10a32c668c551c
We’re playing with nums and signs. signs tell us what to do.
We build numbers as we see them, then use the last sign to decide:
'+' or '-' means put the pre num in the stack.
'*' or '/' means take a num out, mix it with the new num and put it back.